fix: move find job logic to custom hook, import input search
