+++ apps/freelance/src/pages/FindJobs/constants.ts
@@ -2,6 +2,7 @@ export const filterTop = '45px';
 export const filterRight = '0px';
 
 export const fetchLimit = 5;
+export const defaultOffset = 0;
 
 // mock data
 interface User {
